Found Determinant problem and solved
Found problem to be the determinant always came out to be zero if any of the same points were used again. 
Fixed function do amount and rectified to go over a long series of points. This takes a serious amount of time and it's recommended to turn the Step for all of the for loops up to 100 or higher if wanting to check if working